+++ trunk/x10.runtime/src/x10/lang/point.java
@@ -2,7 +2,7 @@ package x10.lang;
 
 abstract public /*value*/ class point/*( region region )*/ extends Object {
 	public final region region;
-	/*parameter*/ public final /*nat*/long rank;
+	/*parameter*/ public final /*nat*/int rank;
 	
 	
 	abstract public static /*value*/ class factory {
@@ -16,16 +16,16 @@ abstract public /*value*/ class point/*( region region )*/ extends Object {
 	}
 	/** Return the value of this point on the i'th dimension.
 	 */    
-	abstract public int valueAt( /*nat*/ long i );
+	abstract public int valueAt( /*nat*/ int i );
 	
 	/** Return true iff the point is on the upper boundary of the i'th
 	 * dimension.
 	 */
-	abstract public boolean onUpperBoundary( /*nat*/ long i );
+	abstract public boolean onUpperBoundary( /*nat*/ int i );
 	
 	/** Return true iff the point is on the lower boundary of the i'th
 	 * dimension.
 	 */
-	abstract public boolean onLowerBoundary( /*nat*/ long i );
+	abstract public boolean onLowerBoundary( /*nat*/ int i );
 	
 }
